requirements.txt
streamlit
python-pptx
pydantic
openai
python-dotenv
